refactor: split chat/thread/security flows and modularize llamacpp utilities #8063
